refactor(condo): DOMA-2567 refactored ticket changes
